Rename FooRequest namespace (#1277)
