+++ src/ArgumentResolver/GitSourceRequestResolver.php
@@ -7,9 +7,9 @@ namespace App\ArgumentResolver;
 use App\Entity\AbstractSource;
 use App\Entity\GitSource;
 use App\Exception\InvalidRequestException;
-use App\FooRequest\Field\StringField;
-use App\FooRequest\StringFieldValidator;
 use App\Request\GitSourceRequest;
+use App\RequestField\Field\StringField;
+use App\RequestField\Validator\StringFieldValidator;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ValueResolverInterface;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
