+++ src/ArgumentResolver/SuiteRequestResolver.php
@@ -8,12 +8,12 @@ use App\Entity\SourceInterface;
 use App\Entity\Suite;
 use App\Exception\EntityNotFoundException;
 use App\Exception\InvalidRequestException;
-use App\FooRequest\Field\StringField;
-use App\FooRequest\Field\YamlFilenameCollectionField;
-use App\FooRequest\StringFieldValidator;
-use App\FooRequest\YamlFilenameCollectionFieldValidator;
 use App\Repository\SourceRepository;
 use App\Request\SuiteRequest;
+use App\RequestField\Field\StringField;
+use App\RequestField\Field\YamlFilenameCollectionField;
+use App\RequestField\Validator\StringFieldValidator;
+use App\RequestField\Validator\YamlFilenameCollectionFieldValidator;
 use App\Security\EntityAccessChecker;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ValueResolverInterface;
