+++ src/Exception/DuplicateFilePathException.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Exception;
 
-use App\FooRequest\Field\Field;
-use App\FooRequest\FieldInterface;
 use App\FooResponse\BadRequestErrorInterface;
+use App\RequestField\Field\Field;
+use App\RequestField\FieldInterface;
 
 class DuplicateFilePathException extends \Exception implements BadRequestErrorInterface
 {
