+++ src/FooResponse/ErrorSerializer/ComponentFactory/FieldFactory.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace App\FooResponse\ErrorSerializer\ComponentFactory;
 
-use App\FooRequest\CollectionFieldInterface;
 use App\FooResponse\BadRequestErrorInterface;
 use App\FooResponse\ErrorInterface;
 use App\FooResponse\ErrorSerializer\Component;
 use App\FooResponse\ErrorSerializer\ComponentFactoryInterface;
+use App\RequestField\CollectionFieldInterface;
 
 class FieldFactory implements ComponentFactoryInterface
 {
