+++ src/FooResponse/ErrorSerializer/ComponentFactory/RequirementsFactory.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace App\FooResponse\ErrorSerializer\ComponentFactory;
 
-use App\FooRequest\RequirementsInterface;
-use App\FooRequest\ScalarRequirementsInterface;
 use App\FooResponse\BadRequestErrorInterface;
 use App\FooResponse\ErrorInterface;
 use App\FooResponse\ErrorSerializer\Component;
 use App\FooResponse\ErrorSerializer\ComponentFactoryInterface;
 use App\FooResponse\SizeInterface;
+use App\RequestField\RequirementsInterface;
+use App\RequestField\ScalarRequirementsInterface;
 
 class RequirementsFactory implements ComponentFactoryInterface
 {
