+++ src/RequestField/RequirementsInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\FooRequest;
+namespace App\RequestField;
 
 interface RequirementsInterface
 {
