+++ src/RequestField/StringFieldInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\FooRequest;
+namespace App\RequestField;
 
 interface StringFieldInterface extends FieldInterface
 {
